moved func to provider
