+++ packages/compass-telemetry/src/index.ts
@@ -5,4 +5,3 @@ export type {
   IdentifyTraits,
   ExtraConnectionData,
 } from './types';
-export { TestName, useFireExperimentViewed } from './growth-experiments';
